bundle up server routes into files and require
