+++ client/app.js
@@ -1,8 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom'
-//import Router from 'react-router';
 import Landing from './components/index.js';
-//let {Route, Link, State, RouteHandler, DefaultRoute} = Router;
 
 class App extends React.Component{
 
@@ -16,15 +14,6 @@ class App extends React.Component{
 
 }
 
-// var routes = (
-//   <RouteHandler handler={App}>
-//     <DefaultRoute handler={Landing} />
-//     <Route name="landing" path="/" handler={Landing} />
-//   </RouteHandler>
-// );
-
-//Router.run(routes, function (Handler) {
-  ReactDOM.render(<App/>,  document.getElementById("app"));
-//});
+ReactDOM.render(<App/>,  document.getElementById("app"));
 
 module.exports = App;
