+++ package.json
@@ -26,7 +26,6 @@
     "react-dom": "^0.14.0",
     "react-router": "^1.0.0",
     "react-tap-event-plugin": "^0.2.1",
-    "request": "^2.61.0",
     "socket.io": "^1.3.7",
     "underscore": "^1.8.3"
   },
