+++ server/db-config.js
@@ -1,0 +1,1 @@
+module.exports = require('mongoskin').db('mongodb://heroku_app31904821:ouldm56b1i98br7lfm5f009olh@ds053370.mongolab.com:53370/heroku_app31904821');
