+++ server/templates/index.html
@@ -27,7 +27,7 @@
     </script>
 
     <script src="/socket.io/socket.io.js"></script>
-    <script src='./app.js'></script>
+    <script src='/app.js'></script>
   </body>
 </html>
